+++ src/main/java/com/github/canbabel/canio/dbc/AttributeDefinitionInt.java
@@ -1,11 +1,11 @@
 package com.github.canbabel.canio.dbc;
 
 class AttributeDefinitionInt extends AttributeDefinition {
-	private final int min;
-	private final int max;
-	private int def;
+	private final long min;
+	private final long max;
+	private long def;
 
-	public AttributeDefinitionInt(String name, AttrTarget target, int min, int max)
+	public AttributeDefinitionInt(String name, AttrTarget target, long min, long max)
 	{
 		super(name, target, AttrType.INT);
 		this.min = min;
@@ -13,22 +13,22 @@ class AttributeDefinitionInt extends AttributeDefinition {
 		def = 0;
 	}
 
-	void setDefault(int def)
+	void setDefault(long def)
 	{
 		this.def = def;
 	}
 	
-	int getDefault()
+	long getDefault()
 	{
 		return def;
 	}
 	
-	int getMax()
+	long getMax()
 	{
 		return max;
 	}
 	
-	int getMin()
+	long getMin()
 	{
 		return min;
 	}
